requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.1.0
drf-nested-routers==0.93.4
